Create Form component and render it in Books
