+++ src/components/Books.js
@@ -1,27 +1,12 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { addBook, removeBook } from '../redux/books/books';
+import { removeBook } from '../redux/books/books';
+import Form from './Form';
 
 const Books = () => {
   const books = useSelector((state) => state.books);
-  const nextIndex = books.length;
   const dispatch = useDispatch();
-  const [bookState, setBookState] = useState({ id: nextIndex });
-  const titleInput = document.getElementById('title');
-  const categoryInput = document.getElementById('category');
 
-  const handleChange = (e) => {
-    setBookState((bookState) => ({
-      ...bookState,
-      [e.target.name]: e.target.value,
-    }));
-  };
-
-  const handleClick = () => {
-    dispatch(addBook({ bookState }));
-    titleInput.value = '';
-    categoryInput.value = '';
-  };
   return (
     <div>
       <ul>
@@ -40,6 +25,7 @@ const Books = () => {
               {book.category}
               {' '}
               -
+              {' '}
               <button
                 type="button"
                 onClick={() => dispatch(removeBook(book))}
@@ -48,19 +34,9 @@ const Books = () => {
               </button>
             </li>
           ))
-}
+        }
       </ul>
-      <form>
-        <h3>Add new book</h3>
-        <input name="title" type="text" placeholder="Book title" id="title" onChange={handleChange} />
-        <input name="category" type="text" placeholder="category" id="category" onChange={handleChange} />
-        <button
-          type="button"
-          onClick={handleClick}
-        >
-          Add book
-        </button>
-      </form>
+      <Form />
     </div>
 
   );
